backend/requirements.txt
# Python dependencies (install with: pip install -r requirements.txt)
fastapi==0.115.0
uvicorn==0.30.0
openai-whisper==20240930
beautifulsoup4==4.12.3
httpx==0.27.0
lxml==5.3.0

# SYSTEM DEPENDENCY (not a Python package):
# ffmpeg — required by Whisper to decode audio files
# Install with: brew install ffmpeg (macOS) or apt install ffmpeg (Linux)
